+++ SparkArtName/src/main/java/key/selectors/CSVTrajIDSelector.java
@@ -6,9 +6,9 @@ import utilities.CSVRecord;
 /**
  * Created by giannis on 11/12/18.
  */
-public class CSVTrajIDSelector implements Function<CSVRecord, Long> {
+public class CSVTrajIDSelector implements Function<CSVRecord, Integer> {
     @Override
-    public Long call(CSVRecord csvRecord) throws Exception {
+    public Integer call(CSVRecord csvRecord) throws Exception {
         return csvRecord.getTrajID();
     }
 }
