+++ SparkArtName/src/main/java/key/selectors/QueryIDSelector.java
@@ -8,9 +8,9 @@ import java.util.Set;
 /**
  * Created by giannis on 12/01/19.
  */
-public class QueryIDSelector implements Function<Tuple2<Long, Set<Long>>, Long> {
+public class QueryIDSelector implements Function<Tuple2<Integer, Set<Integer>>, Integer> {
     @Override
-    public Long call(Tuple2<Long, Set<Long>> v1) throws Exception {
+    public Integer call(Tuple2<Integer, Set<Integer>> v1) throws Exception {
         return v1._1();
     }
 }
