+++ SparkArtName/src/main/java/projections/ProjectRoadSegments.java
@@ -6,9 +6,9 @@ import utilities.CSVRecord;
 /**
  * Created by giannis on 15/01/19.
  */
-public class ProjectRoadSegments implements Function<CSVRecord, Long> {
+public class ProjectRoadSegments implements Function<CSVRecord, Integer> {
     @Override
-    public Long call(CSVRecord v1) throws Exception {
+    public Integer call(CSVRecord v1) throws Exception {
         return v1.getRoadSegment();
     }
 }
